bootstrap: TCP_INFO trace counters, socket latency knobs, Perfetto export
Finalize the deployment-instrumentation branch:
- bootstrap_trace: capture kernel TCP_INFO (rtt/rttvar/rto/cwnd/retrans/lost/
  ca_state) as NetSample records at tcp.ready and after ring_allgather on the
  socket OOB path (recordNetStat / BTRACE_NETSTAT); emitted as BNETSTAT log
  lines and v2 binary records.
- binary dump: append per-flush segments (was truncating, dropping bootstrap
  phases + net samples across the deploy timeline); v2 header adds endian marker,
  format version, and per-thread overflow counts.
- review fixes (rec_26): TLS buffer via unique_ptr (C1), endian marker (C2),
  ring-overflow counters + one-time warn (M2), checked mkdir (M3), acquire
  fast-path in isEnabled (M4), %lu in preload (m1), format versioning (m3).
- socket.cc: opt-in latency knobs NCCL_SOCKET_QUICKACK (set + re-armed per recv),
  NCCL_SOCKET_BUSY_POLL, NCCL_SOCKET_THIN_TIMEOUTS (thin-stream linear RTO).
- merge_bootstrap_trace.py: parse v1/v2 multi-segment binaries + BNETSTAT text,
  --net-csv, --perfetto (phase slices + TCP_INFO counter tracks); 8 unit tests.
- tools: standalone TCP bootstrap-ring reproducer (bootstrap_net_probe.c) with
  cpu-jitter baseline, per-step buckets, delayed-ACK provocation mode, plus
  analysis/run harness scripts.

+++ projects/rccl/tools/scripts/rccl_deploy_trace_preload.c
@@ -54,9 +54,9 @@ static void deploy_log(const char* phase, uint64_t start_ns, uint64_t end_ns, in
   uint64_t dur_us = end_ns > start_ns ? (end_ns - start_ns) / 1000ULL : 0ULL;
   char line[512];
   int n = snprintf(line, sizeof(line),
-      "DEPLOY_TRACE rank=%d pid=%d phase=%s t_ns=%llu dur_us=%llu md=%d bytes=%zu detail=%s\n",
+      "DEPLOY_TRACE rank=%d pid=%d phase=%s t_ns=%llu dur_us=%llu md=%d bytes=%lu detail=%s\n",
       rank, pid, phase, (unsigned long long)start_ns, (unsigned long long)dur_us,
-      md, bytes, detail ? detail : "none");
+      md, (unsigned long)bytes, detail ? detail : "none");
   if (n <= 0) return;
 
   const char* tmpl = getenv("RCCL_DEPLOY_TRACE_FILE");
